requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
peewee==3.17.0
requests==2.31.0
pytz==2023.3